backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
httpx==0.28.1
pydantic-settings==2.8.1
python-multipart==0.0.20
Pillow==11.2.1
pypdf==5.4.0
openpyxl==3.1.5
python-docx==1.1.2
rapidocr-onnxruntime==1.4.4
torch==2.11.0
torchvision==0.26.0
transformers==4.49.0
accelerate==1.6.0
sentencepiece==0.2.0
modelscope==1.35.3
einops==0.8.2
timm==1.0.26
